Time: 28 ms (79.55%), Space: 8.4 MB (93.18%) - LeetHub

+++ 3682-count-the-number-of-arrays-with-k-matching-adjacent-elements/3682-count-the-number-of-arrays-with-k-matching-adjacent-elements.cpp
@@ -7,9 +7,11 @@ public:
         base = base % mod;
 
         while (exp > 0) {
-            if (exp & 1) {
+            if (exp & 1) 
+            {
                 result = (result * base) % mod;
             }
+
             base = (base * base) % mod;
             exp = exp >> 1;
         }
@@ -26,7 +28,7 @@ public:
             den = (den * (r - i)) % mod;
         }
 
-        return (int)((num * power(den, mod - 2)) % mod); 
+        return ( num * power(den,mod-2) ) % mod;
     }
 
     int countGoodArrays(int n, int m, int k) {
